backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
websockets==12.0
openai==1.12.0
python-dotenv==1.0.1
aiofiles==23.2.1
pydantic==2.6.1
